Fixed SoundData:getDuration

+++ src/modules/sound/SoundData.cpp
@@ -147,9 +147,9 @@ int SoundData::getSampleRate() const
 	return sampleRate;
 }
 
-int SoundData::getDuration() const
+float SoundData::getDuration() const
 {
-	return size/(channels*sampleRate*bits/8);
+	return float(size) / (channels*sampleRate*bits/8);
 }
 
 void SoundData::setSample(int i, float sample)
