+++ src/modules/sound/SoundData.h
@@ -49,7 +49,7 @@ public:
 	virtual int getBits() const;
 	virtual int getSampleRate() const;
 
-	virtual int getDuration() const;
+	virtual float getDuration() const;
 
 	void setSample(int i, float sample);
 	float getSample(int i) const;
